game_plus_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.35
psycopg2-binary
asyncpg
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
aiofiles==24.1.0
jinja2==3.1.4
python-multipart==0.0.9
email-validator
google-auth